requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
pydantic-settings
# supabase  <- Removing Supabase
sqlmodel   # <- Adding SQLModel
asyncpg    # <- Adding AsyncPG
telethon
qrcode
email-validator
httpx
python-jose
passlib[bcrypt]
greenlet
slowapi
